pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "random-memorization"
version = "0.1.0"
description = "Random memorization experiments: power-law sampling dynamics on unstructured data"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "numpy",
    "wandb",
    "pyyaml",
    "matplotlib",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
